+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Fixed
 - Fixed capacity and multiplier being set to 1 after being set to 0.
+- Fixed day view showing duration end minutes incorrectly above a certain time [#21]
+
+[#21]: https://github.com/ethercreative/bookings/issues/21 
 
 ## 1.0.0-alpha.15 - 2018-11-14
 ### Added